Delete unused transition files and rename RSS component to RS
